requirements.txt
# default Python packages that does not require installation : logging, pathlib, pickle, re
# external requirements
click==8.1.3
Sphinx==6.1.1
coverage==7.0.3
flake8==6.0.0
python-dotenv==0.21.0
numpy==1.23.4
pandas==1.2.5
pandas-profiling==3.6.1
scikit-learn==1.0.2
scipy==1.9.3
seaborn==0.12.1
shap==0.41.0
sklearn-pandas==1.5.0
dill==0.3.6
joblib==1.2.0
xgboost==1.7.0
catboost==1.1.1
phik==0.12.3
lightgbm==3.3.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
ipython-sql==0.4.1
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-widgets==3.0.5
SQLAlchemy==1.4.46
sqlparse==0.4.3
psycopg2-binary==2.9.5
mysql-connector-python==8.0.31
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
python-dotenv==0.21.0
Flask==2.2.2
setuptools==65.6.3
pip==22.3.1
graphviz==0.20.1
